+++ src/clojisr/v1/impl/renjin/printing.clj
@@ -1,6 +1,8 @@
 (ns clojisr.v1.impl.renjin.printing
   (:require [clojisr.v1.protocols :as prot]))
 
+(set! *warn-on-reflection* true)
+
 (defn print-to-string [session r-obj]
   (->> r-obj
        :object-name