+++ src/clojisr/v1/require.clj
@@ -9,6 +9,8 @@
             [clojisr.v1.impl.java-to-clj :refer [java->clj]]
             [clojure.tools.logging.readable :as log]))
 
+(set! *warn-on-reflection* true)
+
 (defn package-r-object [package-symbol object-symbol]
   (evl/r (format "{%s::`%s`}"
                  (name package-symbol)
@@ -116,4 +118,5 @@
                       {:package-name package-symbol})))))
 
 (defn require-r [& packages]
+  {:deprecated "Please use `require-r` function from `clojisr.v1.r` directly."}
   (run! require-r-package packages))