+++ src/clojisr/v1/using_sessions.clj
@@ -9,6 +9,8 @@
             [clojure.tools.logging.readable :as log])
   (:import clojisr.v1.robject.RObject))
 
+(set! *warn-on-reflection* true)
+
 (defn random-object-name []
   (str mem/session-env "$" (util/rand-name)))